XRDDEV-1544: Define ordering for timestamped records
Define explicit ordering for timestamped records, so that a clustered
setup that shares the messagelog database, all nodes update records
in the same order.

The query returning to-be-timestamped records uses an index that includes
the record id. The "order by id" added by this fix is virtually free
because the index is already sorted by id. Therefore it is also likely
that the records were already returned in the same order and a deadlock
could not occur in practice.