build(serverconf): Correct the `hibernate-tools` dependency
As discovered by an analysis with ORT [1], the `hibernate-tools`
dependency does not exist in the currently set version "6.2.9.Final",
making the `:serverconf:schemaExport` Gradle task fail. It was only
available until version "5.6.15.Final", see [2]. The successor
providing `org.hibernate.tool.ant.HibernateToolTask` seems to be
`hibernate-tools-ant` [3], so use that instead.

However, this does not fully fix running `:serverconf:schemaExport` as
it still fails with

    * What went wrong:
    Execution failed for task ':serverconf:schemaExport'.
    > java.lang.NoClassDefFoundError: ee/ria/xroad/common/identifier/ClientId$Conf

which is something that needs to be investigated separately.

[1]: https://github.com/oss-review-toolkit/ort
[2]: https://mvnrepository.com/artifact/org.hibernate/hibernate-tools
[3]: https://mvnrepository.com/artifact/org.hibernate.tool/hibernate-tools-ant

Signed-off-by: Sebastian Schuberth <sebastian@doubleopen.org>